This adds back the missing `ProposedSchedule` to the server state.
This handles the missing information part in #5436
